Move to unit tests
